backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pgvector==0.3.5

# AI / LLM
anthropic==0.40.0
openai==1.54.0
langgraph==0.2.53
langchain==0.3.7
langchain-openai==0.2.8

# Document parsing
pypdf==5.1.0
beautifulsoup4==4.12.3
httpx==0.27.2

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.6.1
